# public/swagger.yaml
components:
...
    schemas:
...
        mongoose.Types.ObjectId:
            type: string
        CommentResponse:
            properties:
                cid:
                    type: number
                    format: double
                    description: 'Comment ID'
                    example: 1
                commentedBy:
                    $ref: '#/components/schemas/mongoose.Types.ObjectId'
                    description: 'User who made the comment'
                text:
                    type: string
                    description: 'Comment text'
                    example: 'This is a comment'
            required:
                - cid
                - commentedBy
                - text
            type: object
            additionalProperties: false
        PostResponse:
            properties:
                pid:
                    type: number
                    format: double
                    description: 'Post ID'
                    example: 1
                title:
                    type: string
                    description: 'Post title'
                    example: 'Title of the post'
                description:
                    type: string
                    description: 'Post description'
                    example: 'Description of the post'
                likes:
                    items:
                        type: number
                        format: double
                    type: array
                    description: 'Array of user IDs who liked the post'
                    example:
                        - 1
                        - 2
                        - 3
                comments:
                    items:
                        $ref: '#/components/schemas/CommentResponse'
                    type: array
                    description: 'Array of comments on the post'
                postedBy:
                    $ref: '#/components/schemas/mongoose.Types.ObjectId'
            required:
                - pid
                - title
                - description
                - likes
                - comments
                - postedBy
            type: object
            additionalProperties: false
        PostComment:
            properties:
                cid:
                    type: number
                    format: double
                commentedBy:
                    $ref: '#/components/schemas/mongoose.Types.ObjectId'
                text:
                    type: string
            required:
                - cid
                - commentedBy
                - text
            type: object
            additionalProperties: false
        PostPayload:
            properties:
                title:
                    type: string
                    description: 'title of post'
                    example: ""
                description:
                    type: string
                    description: 'description of post'
                    example: ""
                postedBy:
                    $ref: '#/components/schemas/mongoose.Types.ObjectId'
                    description: postedBy
                images:
                    items:
                        type: string
                    type: array
                    description: Image
                    example: ""
                comments:
                    items:
                        $ref: '#/components/schemas/PostComment'
                    type: array
                cost:
                    type: number
                    format: double
                heritage:
                    type: string
                placesToVisit:
                    items:
                        type: string
                    type: array
                communityAccess:
                    type: string
                easeOfTransportation:
                    type: string
                safety:
                    type: string
                location:
                    type: string
            required:
                - title
                - description
                - postedBy
                - cost
                - location
            type: object
            additionalProperties: false
        getPostResponse:
            properties:
                pid:
                    type: number
                    format: double
                title:
                    type: string
                description:
                    type: string
                images:
                    items:
                        type: string
                    type: array
                postedBy:
                    properties:
                        profilePicture:
                            type: string
                        name:
                            type: string
                        id:
                            type: number
                            format: double
                        _id:
                            type: string
                    required:
                        - profilePicture
                        - name
                        - id
                        - _id
                    type: object
                likes:
                    items:
                        $ref: '#/components/schemas/mongoose.Types.ObjectId'
                    type: array
                liked:
                    type: boolean
            required:
                - pid
                - title
                - description
                - postedBy
                - likes
            type: object
            additionalProperties: false
        paginationResponse:
            properties:
                page:
                    type: number
                    format: double
                limit:
                    type: number
                    format: double
                totalPages:
                    type: number
                    format: double
                totalPosts:
                    type: number
                    format: double
                data:
                    items:
                        $ref: '#/components/schemas/getPostResponse'
                    type: array
            required:
                - page
                - limit
                - totalPages
                - totalPosts
                - data
            type: object
            additionalProperties: false
        GetDetailsResponse:
            properties:
                location:
                    type: string
                heritage:
                    type: string
                placesToVisit:
                    items:
                        type: string
                    type: array
                communityAccess:
                    type: string
                easeOfTransportation:
                    type: string
                safety:
                    type: string
                cost:
                    type: number
                    format: double
            required:
                - location
                - cost
            type: object
            additionalProperties: false
        Like:
            properties:
                _id:
                    type: string
                id:
                    type: number
                    format: double
                name:
                    type: string
                profilePicture:
                    type: string
            required:
                - _id
                - id
                - name
                - profilePicture
            type: object
            additionalProperties: false
        LikesResponse:
            properties:
                count:
                    type: number
                    format: double
                data:
                    items:
                        $ref: '#/components/schemas/Like'
                    type: array
                    nullable: true
            required:
                - count
                - data
            type: object
            additionalProperties: false
        getCommentResponse:
            properties:
                cid:
                    type: number
                    format: double
                commentedBy:
                    properties:
                        profilePicture:
                            type: string
                        name:
                            type: string
                        id:
                            type: number
                            format: double
                    required:
                        - profilePicture
                        - name
                        - id
                    type: object
                text:
                    type: string
            required:
                - cid
                - commentedBy
                - text
            type: object
            additionalProperties: false
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
info:
    title: express-ts
    version: 1.0.0
    license:
        name: ISC
    contact: {}
openapi: 3.0.0
paths:
...
    /posts/createpost:
        post:
            operationId: CreatePost
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                anyOf:
                                    - {$ref: '#/components/schemas/PostResponse'}
                                    - {type: string}
            tags:
                - Post
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PostPayload'
    /posts/search:
        post:
            operationId: Search
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/getPostResponse'
                                type: array
            tags:
                - Post
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                page:
                                    type: number
                                    format: double
                                query:
                                    type: string
                            required:
                                - page
                                - query
                            type: object
...
    /posts/getuserposts:
        get:
            operationId: FindPostsByUserId
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/paginationResponse'
            tags:
                - Post
            security: []
            parameters:
                -
                    in: query
                    name: userId
                    required: true
                    schema:
                        format: double
                        type: number
    /posts/getdetails:
        get:
            operationId: Getdetails
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetDetailsResponse'
            tags:
                - Post
            security: []
            parameters:
                -
                    in: query
                    name: postId
                    required: true
                    schema:
                        format: double
                        type: number
    /posts/delete:
        post:
            operationId: DeletePost
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                type: string
            tags:
                - Post
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                pid:
                                    type: number
                                    format: double
                            required:
                                - pid
                            type: object
    /posts/like:
        put:
            operationId: LikePost
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                anyOf:
                                    - {$ref: '#/components/schemas/PostResponse'}
                                    - {type: string}
            tags:
                - Post
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                userId:
                                    type: number
                                    format: double
                                pid:
                                    type: number
                                    format: double
                            required:
                                - userId
                                - pid
                            type: object
    /posts/likescount:
        get:
            operationId: GetLikesCount
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                type: number
                                format: double
            tags:
                - Post
            security: []
            parameters:
                -
                    in: query
                    name: postId
                    required: true
                    schema:
                        format: double
                        type: number
    /posts/getlikes:
        get:
            operationId: GetLikes
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/LikesResponse'
                                type: array
            tags:
                - Post
            security: []
            parameters:
                -
                    in: query
                    name: postId
                    required: true
                    schema:
                        format: double
                        type: number
    /posts/getcomments:
        get:
            operationId: GetComments
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/getCommentResponse'
                                type: array
            tags:
                - Post
            security: []
            parameters:
                -
                    in: query
                    name: postId
                    required: true
                    schema:
                        format: double
                        type: number
    /posts/addComment:
        post:
            operationId: AddComment
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CommentResponse'
            tags:
                - Post
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                userId:
                                    type: number
                                    format: double
                                text:
                                    type: string
                                pid:
                                    type: number
                                    format: double
                            required:
                                - userId
                                - text
                                - pid
                            type: object
    '/posts/editcomment/{cid}':
        put:
            operationId: EditComment
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CommentResponse'
            tags:
                - Post
            security: []
            parameters:
                -
                    in: path
                    name: cid
                    required: true
                    schema:
                        format: double
                        type: number
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                text:
                                    type: string
                            required:
                                - text
                            type: object
    '/posts/deletecomment/{cid}':
        delete:
            operationId: DeleteComment
            responses:
                '204':
                    description: 'No content'
            tags:
                - Post
            security: []
            parameters:
                -
                    in: path
                    name: cid
                    required: true
                    schema:
                        format: double
                        type: number
servers:
    -
        url: /
...